Add textfields documentation
- split demo for inputField and textArea
- add documentation page for each of the two
